requirements.txt
httplib2==0.9.2
Django==1.11.4
six==1.10.0
django_crispy_forms==1.7.2
django_formtools==2.1
future==0.16.0
django_reversion==2.0.13
xlsxwriter==1.0.2
xlwt==1.3.0
requests==2.18.4
django-import-export==1.0.0
django-simple-captcha==0.5.6
django-pure-pagination==0.3.0